refactor(clipboard): group active-clipboard files into subdirectories
Collapse the flat `active_clipboard_*` / `active_state_*` module-prefix
layout into proper subdirectories so each layer's active-clipboard files
live together instead of being scattered with a name prefix:

- uc-core:        ports/clipboard/active_clipboard/{dispatch,pull,receiver,register}.rs
- uc-infra:       network/iroh/active_clipboard/{wire,dispatch_adapter,receiver_adapter,pull_wire,pull_client_adapter,pull_serve_adapter}.rs
- uc-application: usecases/clipboard_sync/active_state/{fanout,reconcile,apply_inbound,restore_broadcast_worker,peer_online_resync_worker,serve_pull}.rs

The shared gating/snapshot helpers (send_gate, receive_gate,
snapshot_from_entry) deliberately stay in clipboard_sync/ since the
generic inbound and resend use cases depend on them too.

Pure move plus import-path updates; no behavior change. Top-level
re-exports (uc_core::ports, iroh adapters) are unchanged, so no external
consumer is affected.
