+++ crates/uc-application/src/usecases/clipboard_sync/mod.rs
@@ -10,9 +10,8 @@
 //! consumers (daemon / Tauri / CLI) reach them through
 //! `ClipboardSyncFacade`.
 
-pub(crate) mod active_state_fanout;
+pub(crate) mod active_state;
 pub(crate) mod apply_inbound;
-pub(crate) mod apply_inbound_active_state;
 pub(crate) mod dispatch_entry;
 pub(crate) mod get_entry_delivery_view;
 pub(crate) mod ingest_inbound;
@@ -21,13 +20,9 @@ pub(crate) mod ingest_inbound;
 /// Individual private helpers inside stay scoped via their own
 /// `pub(crate)` / no-modifier visibility.
 pub mod payload_codec;
-pub(crate) mod peer_online_resync_worker;
 pub(crate) mod receive_gate;
-pub(crate) mod reconcile_active_state;
 pub(crate) mod resend_entry;
-pub(crate) mod restore_broadcast_worker;
 pub(crate) mod send_gate;
-pub(crate) mod serve_pull;
 pub(crate) mod snapshot_from_entry;
 
 pub(crate) use dispatch_entry::{
