+++ crates/uc-infra/src/network/iroh/mod.rs
@@ -4,12 +4,7 @@
 //! endpoint lifecycle, session opener, blob transfer. Slice 1 only ships
 //! [`IrohIdentityStore`]; later slices add the rest.
 
-pub mod active_clipboard_dispatch_adapter;
-pub mod active_clipboard_pull_client_adapter;
-pub mod active_clipboard_pull_serve_adapter;
-pub mod active_clipboard_pull_wire;
-pub mod active_clipboard_receiver_adapter;
-pub mod active_clipboard_wire;
+pub mod active_clipboard;
 mod addr_filter;
 pub mod blobs;
 pub mod clipboard_dispatch_adapter;
@@ -26,14 +21,11 @@ pub(crate) mod runtime_consts;
 pub mod transfer_progress_adapter;
 pub mod transfer_progress_wire;
 
-pub use active_clipboard_dispatch_adapter::IrohActiveClipboardDispatchAdapter;
-pub use active_clipboard_pull_client_adapter::IrohActiveClipboardPullClientAdapter;
-pub use active_clipboard_pull_serve_adapter::{
+pub use active_clipboard::{
+    IrohActiveClipboardDispatchAdapter, IrohActiveClipboardPullClientAdapter,
     IrohActiveClipboardPullServeAdapter, IrohActiveClipboardPullServeHandler,
-    ACTIVE_CLIPBOARD_PULL_ALPN,
-};
-pub use active_clipboard_receiver_adapter::{
     IrohActiveClipboardReceiverAdapter, IrohActiveClipboardReceiverHandler, ACTIVE_CLIPBOARD_ALPN,
+    ACTIVE_CLIPBOARD_PULL_ALPN,
 };
 pub(crate) use addr_filter::filter_endpoint_addr;
 pub use blobs::{IrohBlobTransferAdapter, BLOBS_ALPN};
